pyproject.toml
[project]
name = "ai-web-studio"
version = "0.1.0"
description = ""
authors = [
    {name = "MaxBakshaev",email = "risen.2011@mail.ru"}
]
readme = "README.md"
requires-python = "3.12.10"
dependencies = [
    "fastapi (>=0.124.4,<0.125.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "bcrypt (<4)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.22.0"

